refactor: id타입을 Long으로 변경

+++ src/main/java/racingcar/dao/CarResultDao.java
@@ -5,9 +5,9 @@ import racingcar.domain.CarResult;
 import java.util.List;
 
 public interface CarResultDao {
-    int save(CarResult carResult);
+    long save(CarResult carResult);
 
-    CarResult findById(int id);
+    CarResult findById(long id);
 
-    List<CarResult> findAllByPlayResultId(int playResultId);
+    List<CarResult> findAllByPlayResultId(long playResultId);
 }
