+++ src/test/java/racingcar/dao/JdbcGameResultDaoTest.java
@@ -34,7 +34,7 @@ class JdbcGameResultDaoTest {
     @Test
     void save_메서드는_데이터베이스에_게임결과를_저장() {
         GameResult gameResult = new GameResult(10, "hongo", Timestamp.valueOf(LocalDateTime.now()));
-        int id = webPlayResultDao.save(gameResult);
+        long id = webPlayResultDao.save(gameResult);
         GameResult result = webPlayResultDao.findById(id);
         assertThat(result).isNotNull();
     }
@@ -42,7 +42,7 @@ class JdbcGameResultDaoTest {
     @Test
     void findById_메서드는_특정Id의_게임결과를_반환() {
         GameResult gameResult = new GameResult(10, "hongo", Timestamp.valueOf("2023-04-19 00:00:00"));
-        int id = webPlayResultDao.save(gameResult);
+        long id = webPlayResultDao.save(gameResult);
         GameResult savedResult = webPlayResultDao.findById(id);
         assertAll(
                 () -> assertThat(savedResult.getId()).isEqualTo(id),
@@ -56,8 +56,8 @@ class JdbcGameResultDaoTest {
     void findAll_메서드는_모든_게임결과를_반환() {
         GameResult gameResult1 = new GameResult(10, "juno", Timestamp.valueOf("2023-04-19 00:00:00"));
         GameResult gameResult2 = new GameResult(15, "hongo", Timestamp.valueOf("2023-04-19 00:00:00"));
-        int playResult1Id = webPlayResultDao.save(gameResult1);
-        int playResult2Id = webPlayResultDao.save(gameResult2);
+        long playResult1Id = webPlayResultDao.save(gameResult1);
+        long playResult2Id = webPlayResultDao.save(gameResult2);
 
         List<GameResult> savedGameResults = webPlayResultDao.findAll();
         assertAll(
